feat: refractor to enable non breaking change
